requirements.txt
torch
torchvision
transformers 
diffusers 
accelerate
Pillow
matplotlib
fastapi
uvicorn
beautifulsoup4
lxml
html5lib
selenium==4.9.0